pyproject.toml
[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "instruction-learning"
version = "0.1.0"
description = "Lightweight instruction-following embedding adaptor."
license = "MIT"
readme = "README.md"
authors = ["Kirill Borodin"]
keywords = ["nlp", "embeddings", "hydra", "pytorch", "mlops"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
packages = [{ include = "instruction_learning", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
torch = ">=2.5.0,<3.0"
lightning = ">=2.4.0,<3.0"
hydra-core = ">=1.3.2"
mlflow = ">=2.16"
numpy = ">=1.26"
pandas = ">=2.2"
scikit-learn = ">=1.5"
dvc = ">=3.50"
fire = ">=0.6"
onnx = ">=1.16"
onnxruntime = ">=1.18"
rich = ">=13.9"
transformers = ">=4.44"
sentencepiece = ">=0.2"
tqdm = ">=4.66"
sentence-transformers = ">=3.0"
datasets = ">=3.0"

[tool.poetry.group.dev.dependencies]
black = ">=24.4"
isort = ">=5.13"
flake8 = ">=7.1"
pre-commit = ">=4.0"
pytest = ">=8.3"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88
known_first_party = ["instruction_learning"]

[tool.flake8]
max-line-length = 88
extend-ignore = ["E203", "W503"]
per-file-ignores = ["__init__.py:F401"]

[tool.pytest.ini_options]
minversion = "8.3"
addopts = "-ra"
testpaths = ["tests"]
